[FIX] the update and delete user, and fix the delete post

+++ groupomania/back/controllers/post.js
@@ -64,11 +64,11 @@ exports.getOne = (req, res, next) => {
 
 exports.deleteOne = (req, res, next) => {
   Post.findOne({where:{ id: req.params.id }})
-    .then(Post => {
-      if (req.userId === Post.userId || req.isAdmin) {
+    .then(post => {
+      if (req.userId === post.userId || req.isAdmin) {
         //if post.image differente of null, delete the image
-        if (Post.image !== null) {
-          const imageName = Post.image.split('/images/')[1];
+        if (post.image !== '') {
+          const imageName = post.image.split('/images/')[1];
           fs.unlink(`images/${imageName}`, (error) => {
             if (error) {
               return res.status(500).json({ error: 'Erreur lors de la suppression de l\'image' });
@@ -93,26 +93,27 @@ exports.deleteOne = (req, res, next) => {
 
 exports.updateOne = (req, res, next) => {
   Post.findOne({where:{ id: req.params.id }})
-    .then(Post => {
-      if (req.userId === Post.userId || req.isAdmin) {
+    .then(post => {
+      if (req.userId === post.userId || req.isAdmin) {
         //if there is a file, used multer to delete the old file and save the new
-        if (req.file) {
-          const imageName = Post.image.split('/images/')[1];
-          fs.unlink(`images/${imageName}`, (error) => {
-            if (error) {
-              return res.status(500).json({ error: 'Erreur lors de la suppression de l\'image' });
-            }
-          })    
+        if (post.image !== '' && req.file) {
+          const imageName = post.image.split('/images/')[1];
+            fs.unlink(`images/${imageName}`, (error) => {
+              if (error) {
+                res.status(500).json({ error: 'Erreur lors de la suppression de l\'image' });
+                return
+              }
+            })  
         }
         const temp = JSON.stringify(req.body);
         const postObject = JSON.parse(temp);
         //ternaire operator to check if there is a file
-        const post = req.file ? {
+        const postItem = req.file ? {
           //using decomposition to copy the property of an source object on a new object
           ...postObject,
           image: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
         } : { ...postObject};
-        Post.update(post, {where:{ id: req.params.id }})
+        Post.update(postItem, {where:{ id: req.params.id }})
           .then(() => res.status(200).json({ message: 'Post modifié !' }))
           .catch(error => res.status(400).json({ error }));
       } else {
